requirements.txt
# 65
transformers
torch
Pillow
sentencepiece
accelerate
protobuf
bitsandbytes
python-docx
lingua-language-detector
python-dotenv
customtkinter
